remove redundent code
