+++ electricpy/__init__.py
@@ -22,6 +22,7 @@ __version__ = _version_  # Alias Version for User Ease
 
 # Import Submodules
 from .constants import *
+from .phasor import compose, parallelz
 
 # Import Supporting Modules
 import numpy as _np
@@ -34,494 +35,6 @@ from inspect import getframeinfo as _getframeinfo
 from inspect import stack as _stack
 from scipy.integrate import quad as integrate
 import scipy.signal as sig
-
-
-# Define Phase Angle Generator
-def phs(ang):
-    """
-    Complex Phase Angle Generator.
-
-    Generate a complex value given the phase angle
-    for the complex value.
-
-    Same as `phase`.
-
-    Parameters
-    ----------
-    ang:        float
-                The angle (in degrees) for which
-                the value should be calculated.
-
-    See Also
-    --------
-    phasorlist: Phasor Generator for List or Array
-    cprint:     Complex Variable Printing Function
-    phasorz:    Impedance Phasor Generator
-    phasor:     Phasor Generating Function
-    """
-    # Return the Complex Angle Modulator
-    return (_np.exp(1j * _np.radians(ang)))
-
-
-phase = phs  # Create Duplicate Name
-
-
-# Define Phasor Generator
-def phasor(mag, ang=0):
-    """
-    Complex Phasor Generator.
-
-    Generates the standard Pythonic complex representation
-    of a phasor voltage or current when given the magnitude
-    and angle of the specific voltage or current.
-
-    Parameters
-    ----------
-    mag:        float
-                The Magnitude of the Voltage/Current
-    ang:        float
-                The Angle (in degrees) of the Voltage/Current
-
-    Returns
-    -------
-    phasor:     complex
-                Standard Pythonic Complex Representation of
-                the specified voltage or current.
-
-    Examples
-    --------
-    >>> import electricpy as ep
-    >>> ep.phasor(67, 120) # 67 volts at angle 120 degrees
-    (-33.499999999999986+58.02370205355739j)
-
-    See Also
-    --------
-    phasorlist: Phasor Generator for List or Array
-    cprint:     Complex Variable Printing Function
-    phasorz:    Impedance Phasor Generator
-    phs:        Complex Phase Angle Generator
-    """
-    # Test for Tuple/List Arg
-    if isinstance(mag, (tuple, list, _np.ndarray)):
-        ang = mag[1]
-        mag = mag[0]
-    return (_c.rect(mag, _np.radians(ang)))
-
-
-# Define Phasor Array Generator
-def phasorlist(arr):
-    """
-    Complex Phasor Generator for 2-D Array or 2-D List.
-
-    Generates the standard Pythonic complex representation
-    of a phasor voltage or current when given the magnitude
-    and angle of the specific voltage or current for a list
-    or array of values.
-
-    Parameters
-    ----------
-    arr:        array-like
-                2-D array or list of magnitudes and angles.
-                Each item must be set of magnitude and angle
-                in form of: [mag, ang].
-
-    Returns
-    -------
-    phasor:     complex
-                Standard Pythonic Complex Representation of
-                the specified voltage or current.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> import electricpy as ep
-    >>> voltages = np.array([[67,0],
-                             [67,-120],
-                             [67,120]])
-    >>> Vset = ep.phasorlist( voltages )
-    >>> print(Vset)
-
-    See Also
-    --------
-    phasor:     Phasor Generating Function
-    vectarray:  Magnitude/Angle Array Pairing Function
-    cprint:     Complex Variable Printing Function
-    phasorz:    Impedance Phasor Generator
-    """
-    # Use List Comprehension to Process
-
-    # Return Array
-    return _np.array([phasor(i) for i in arr])
-
-
-# Define Vector Array Generator
-def vectarray(arr, degrees=True, flatarray=False):
-    """
-    Format Complex as Array of Magnitude/Angle Pairs.
-
-    Consume an iterable (list/tuple/ndarray/etc.) of
-    complex numbers and generate an ndarray of magnitude
-    and angle pairs, formatted as a 2-dimension (or
-    optionally 1-dimension) array.
-
-    Parameters
-    ----------
-    arr:        array-like
-                Array or list of complex numbers to be
-                cast to magnitude/angle pairs.
-    degrees:    bool, optional
-                Control option to set the angles in
-                degrees. Defaults to True.
-    flatarray:  bool, optional
-                Control option to set the array return
-                to work as a 1-dimension list. Defaults
-                to False, formatting as a 2-dimension
-                list.
-
-    Returns
-    -------
-    polararr:   ndarray
-                Array of magnitude/angle pairs as a
-                2-dimension array (or optionally
-                1-dimension array).
-
-    See Also
-    --------
-    phasor:     Phasor Generating Function
-    phasorlist: Phasor Generator for List or Array
-    """
-    # Iteratively Append Arrays to the Base
-
-    def vector_cast(num):
-        mag, ang = _c.polar(num)
-
-        if degrees:
-            ang = _np.degrees(ang)
-
-        return [mag, ang]
-
-    polararr = _np.array([vector_cast(num) for num in arr])
-    # Reshape Array if Needed
-    if not flatarray:
-        polararr = _np.reshape(polararr, (-1, 2))
-    return polararr
-
-
-# Define Phasor Data Generator
-def phasordata(mn, mx=None, npts=1000, mag=1, ang=0, freq=60,
-               retstep=False, rettime=False, sine=False):
-    """
-    Complex Phasor Data Generator.
-
-    Generates a sinusoidal data set with minimum, maximum,
-    frequency, magnitude, and phase angle arguments.
-
-    Parameters
-    ----------
-    mn:         float, optional
-                Minimum time (in seconds) to generate data for.
-                default=0
-    mx:         float
-                Maximum time (in seconds) to generate data for.
-    npts:       float, optional
-                Number of data samples. default=1000
-    mag:        float, optional
-                Sinusoid magnitude, default=1
-    ang:        float, optional
-                Sinusoid angle in degrees, default=0
-    freq:       float, optional
-                Sinusoid frequency in Hz
-    retstep:    bool, optional
-                Control argument to request return of time
-                step size (dt) in seconds.
-    sine:       bool, optional
-                Control argument to require data be generated
-                with a sinusoidal function instead of cosine.
-
-    Returns
-    -------
-    data:       numpy.ndarray
-                The resultant data array.
-    """
-    # Test Inputs for Min/Max
-    if mx == None:
-        # No Minimum provided, use Value as Maximum
-        mx = mn
-        mn = 0
-    # Generate Omega
-    w = 2 * _np.pi * freq
-    # Generate Time Array
-    t, dt = _np.linspace(mn, mx, npts, retstep=True)
-    # Generate Data Array
-    if not sine:
-        data = mag * _np.cos(w * t + _np.radians(ang))
-    else:
-        data = mag * _np.sin(w * t + _np.radians(ang))
-    # Generate Return Data Set
-    dataset = [data]
-    if retstep:
-        dataset.append(dt)
-    if rettime:
-        dataset.append(t)
-    # Return Dataset
-    if len(dataset) == 1:
-        return (dataset[0])
-    return (dataset)
-
-
-# Define Complex LaTeX Generator
-def clatex(val, round=3, polar=True, predollar=True, postdollar=True,
-           double=False):
-    """
-    Complex Value Latex Generator.
-
-    Function to generate a LaTeX string of complex value(s)
-    in either polar or rectangular form. May generate both dollar
-    signs.
-
-    Parameters
-    ----------
-    val:        complex
-                The complex value to be printed, if value
-                is a list or numpy array, the result will be
-                demonstrated as a matrix.
-    round:      int, optional
-                Control to specify number of decimal places
-                that should displayed. default=True
-    polar:      bool, optional
-                Control argument to force result into polar
-                coordinates instead of rectangular. default=True
-    predollar:  bool, optional
-                Control argument to enable/disable the dollar
-                sign before the string. default=True
-    postdollar: bool, optional
-                Control argument to enable/disable the dollar
-                sign after the string. default=True
-    double:     bool, optional
-                Control argument to specify whether or not
-                LaTeX dollar signs should be double or single,
-                default=False
-
-    Returns
-    -------
-    latex:      str
-                LaTeX string for the complex value.
-    """
-    # Define Interpretation Functions
-    def polarstring(val, round):
-        mag, ang_r = _c.polar(val)  # Convert to polar form
-        ang = _np.degrees(ang_r)  # Convert to degrees
-        mag = _np.around(mag, round)  # Round
-        ang = _np.around(ang, round)  # Round
-        latex = str(mag) + '∠' + str(ang) + '°'
-        return (latex)
-
-    def rectstring(val, round):
-        real = _np.around(val.real, round)  # Round
-        imag = _np.around(val.imag, round)  # Round
-        if imag > 0:
-            latex = str(real) + "+j" + str(imag)
-        else:
-            latex = str(real) + "-j" + str(abs(imag))
-        return (latex)
-
-    # Interpret as numpy array if simple list
-    if isinstance(val, list):
-        val = _np.asarray(val)  # Ensure that input is array
-    # Find length of the input array
-    if isinstance(val, _np.ndarray):
-        shp = val.shape
-        try:
-            row, col = shp  # Interpret Shape of Object
-        except:
-            row = shp[0]
-            col = 1
-        sz = val.size
-        # Open Matrix
-        latex = r'\begin{bmatrix}'
-        # Iteratively Process Each Item in Array
-        for ri in range(row):
-            if ri != 0:  # Insert Row Separator
-                latex += r'\\'
-            if col > 1:
-                for ci in range(col):
-                    if ci != 0:  # Insert Column Separator
-                        latex += r' & '
-                    # Add Complex Represetation of Value
-                    if polar:
-                        latex += polarstring(val[ri][ci], round)
-                    else:
-                        latex += rectstring(val[ri][ci], round)
-            else:
-                # Add Complex Represetation of Value
-                if polar:
-                    latex += polarstring(val[ri], round)
-                else:
-                    latex += rectstring(val[ri], round)
-        # Close Matrix
-        latex += r'\end{bmatrix}'
-    elif isinstance(val, complex):
-        # Treat as Polar When Directed
-        if polar:
-            latex = polarstring(val, round)
-        else:
-            latex = rectstring(val, round)
-    else:
-        raise ValueError("Invalid Input Type")
-    # Add Dollar Sign pre-post
-    if double:
-        dollar = r'$$'
-    else:
-        dollar = r'$'
-    if predollar:
-        latex = dollar + latex
-    if postdollar:
-        latex = latex + dollar
-    return (latex)
-
-
-# Define Transfer Function LaTeX Generator
-def tflatex(sys, sysp=None, var='s', predollar=True,
-            postdollar=True, double=False, tolerance=1e-8):
-    r"""
-    Transfer Function LaTeX String Generator.
-
-    LaTeX string generating function to create a transfer
-    function string in LaTeX. Particularly useful for
-    demonstrating systems in Interactive Python Notebooks.
-
-    Parameters
-    ----------
-    sys:        list
-                If provided in conjunction with optional
-                parameter `sysp`, the parameter `sys` will
-                act as the numerator set. Otherwise, can be
-                passed as a list containing two sublists,
-                the first being the numerator set, and the
-                second being the denominator set.
-    sysp:       list, optional
-                If provided, this input will act as the
-                denominator of the transfer function.
-    var:        str, optional
-                The variable that should be printed for each
-                term (i.e. 's' or 'j\omega'). default='s'
-    predollar:  bool, optional
-                Control argument to enable/disable the dollar
-                sign before the string. default=True
-    postdollar: bool, optional
-                Control argument to enable/disable the dollar
-                sign after the string. default=True
-    double:     bool, optional
-                Control argument to specify whether or not
-                LaTeX dollar signs should be double or single,
-                default=False
-    tolerance:  float, optional
-                The floating point tolerance cutoff to evaluate
-                each term against. If the absolute value of the
-                particular term is greater than the tolerance,
-                the value will be printed, if not, it will not
-                be printed. default=1e-8
-
-    Returns
-    -------
-    latex:      str
-                LaTeX string for the transfer function.
-    """
-    # Collect Numerator and Denominator Terms
-    if isinstance(sysp, (list, tuple, _np.ndarray)):
-        num = sys
-        den = sysp
-    else:
-        num, den = sys
-
-    # Generate String Function
-    def genstring(val):
-        length = len(val)
-        strg = ''
-        for i, v in enumerate(val):
-            # Add Each Term to String
-            if abs(v) > tolerance:
-                # Add '+' Symbol After Each Term
-                if i != 0:
-                    strg += r'+'
-                strg += str(v)
-                # Determine Exponent
-                xpnt = length - i - 1
-                if xpnt == 1:
-                    strg += var
-                elif xpnt == 0:
-                    pass  # Don't Do Anything
-                else:
-                    strg += var + r'^{' + str(xpnt) + r'}'
-        return (strg)
-
-    # Generate Total TF String
-    latex = r'\frac{' + genstring(num) + r'}{'
-    latex += genstring(den) + r'}'
-    # Add Dollar Sign pre-post
-    if double:
-        dollar = r'$$'
-    else:
-        dollar = r'$'
-    if predollar:
-        latex = dollar + latex
-    if postdollar:
-        latex = latex + dollar
-    return (latex)
-
-
-# Define Complex Composition Function
-def compose(*arr):
-    """
-    Complex Composition Function.
-
-    Accepts a set of real values and generates an array
-    of complex values. Input must be array-like, but can
-    appear in various forms:
-
-    - [ real, imag]
-    - [ [ real1, ..., realn ], [ imag1, ..., imagn ] ]
-    - [ [ real1, imag1 ], ..., [ realn, imagn ] ]
-
-    Will always return values in form:
-
-    [ complex1, ... complexn ]
-
-    Parameters
-    ----------
-    arr:        array_like
-                The input of real and imaginary term(s)
-    """
-    # Condition Input
-    if len(arr) == 1:
-        arr = arr[0]  # Extract 0-th term
-    # Input comes in various forms, we must first detect shape
-    arr = _np.asarray(arr)  # Format as Numpy Array
-    # Gather Shape to Detect Format
-    try:
-        row, col = arr.shape
-        # Passed Test, Valid Shape
-        retarr = _np.array([])  # Empty Return Array
-        # Now, Determine whether is type 2 or 3
-        if col == 2:  # Type 3
-            for i in range(row):  # Iterate over each row
-                item = arr[i][0] + 1j * arr[i][1]
-                retarr = _np.append(retarr, item)
-        elif row == 2:  # Type 2
-            for i in range(col):  # Iterate over each column
-                item = arr[0][i] + 1j * arr[1][i]
-                retarr = _np.append(retarr, item)
-        else:
-            raise ValueError("Invalid Array Shape, must be 2xN or Nx2.")
-        # Successfully Generated Array, Return
-        return (retarr)
-    except:  # 1-Dimension Array
-        length = arr.size
-        # Test for invalid Array Size
-        if length != 2:
-            raise ValueError("Invalid Array Size, Saw Length of " + str(length))
-        # Valid Size, Calculate and Return
-        return (arr[0] + 1j * arr[1])
 
 
 # Define Cycle Time Function
@@ -890,58 +403,6 @@ def phasorz(C=None, L=None, freq=60, complex=True):
     return (Z)
 
 
-# Define Parallel Impedance Adder
-def parallelz(*args):
-    r"""
-    Parallel Impedance Calculator.
-
-    This function is designed to generate the total parallel
-    impedance of a set (tuple) of impedances specified as real
-    or complex values.
-
-    .. math::
-       Z_{eq}=(\frac{1}{Z_1}+\frac{1}{Z_2}+\dots+\frac{1}{Z_n})^{-1}
-
-    Parameters
-    ----------
-    Z:      tuple of complex
-            The tupled input set of impedances, may be a tuple
-            of any size greater than 2. May be real, complex, or
-            a combination of the two.
-
-    Returns
-    -------
-    Zp:     complex
-            The calculated parallel impedance of the input tuple.
-    """
-    # Gather length (number of elements in tuple)
-    L = len(args)
-    if L == 1:
-        Z = args[0]  # Only One Tuple Provided
-        try:
-            L = len(Z)
-            if (L == 1):
-                Zp = Z[0]  # Only one impedance, burried in tuple
-            else:
-                # Inversely add the first two elements in tuple
-                Zp = (1 / Z[0] + 1 / Z[1]) ** (-1)
-                # If there are more than two elements, add them all inversely
-                if (L > 2):
-                    for i in range(2, L):
-                        Zp = (1 / Zp + 1 / Z[i]) ** (-1)
-        except:
-            Zp = Z  # Only one impedance
-    else:
-        Z = args  # Set of Args acts as Tuple
-        # Inversely add the first two elements in tuple
-        Zp = (1 / Z[0] + 1 / Z[1]) ** (-1)
-        # If there are more than two elements, add them all inversely
-        if (L > 2):
-            for i in range(2, L):
-                Zp = (1 / Zp + 1 / Z[i]) ** (-1)
-    return (Zp)
-
-
 # Define Phase/Line Converter
 def phaseline(VLL=None, VLN=None, Iline=None, Iphase=None, realonly=None, **kwargs):
     r"""
@@ -1301,119 +762,6 @@ def transformertest(Poc=False, Voc=False, Ioc=False, Psc=False, Vsc=False,
               "Not enough arguments were provided.")
 
 
-# Define Phasor Plot Generator
-def phasorplot(phasor, title="Phasor Diagram", legend=False, bg=None,
-               colors=None, radius=None, linewidth=None, size=None,
-               filename=None, plot=True, label=False, labels=False,
-               tolerance=None):
-    """
-    Phasor Plotting Function.
-
-    This function is designed to plot a phasor-diagram with angles in degrees
-    for up to 12 phasor sets. Phasors must be passed as a complex number set,
-    (e.g. [ m+ja, m+ja, m+ja, ... , m+ja ] ).
-
-    Parameters
-    ----------
-    phasor:     list of complex
-                The set of phasors to be plotted.
-    title:      string, optional
-                The Plot Title, default="Phasor Diagram"
-    legend:     bool, optional
-                Control argument to enable displaying the legend, must be passed
-                as an array or list of strings. `label` and `labels` are mimic-
-                arguments and will perform similar operation, default=False
-    bg:         string, optional
-                Background-Color control, default="#d5de9c"
-    radius:     float, optional
-                The diagram radius, unless specified, automatically scales
-    colors:     list of str, optional
-                List of hexidecimal color strings denoting the line colors to use.
-    filename:   string, optional
-                String of filename, if set, will force function to save image.
-    plot:       bool, optional
-                Control argument to disable plotting. default=True
-    size:       float, optional
-                Control argument for figure size. default=None
-    linewidth:  float, optional
-                Control argument to declare the line thickness. default=None
-    tolerance:  float, optional
-                Minimum magnitude to plot, anything less than tolerance will be
-                plotted as a single point at the origin, by default, the tolerance
-                is scaled to be 1/25-th the maximum radius. To disable the tolerance,
-                simply provide either False or -1.
-    """
-    # Load Complex Values if Necessary
-    try:
-        len(phasor)
-    except TypeError:
-        phasor = [phasor]
-    # Manage Colors
-    if colors == None:
-        colors = ["#FF0000", "#800000", "#FFFF00", "#808000", "#00ff00", "#008000",
-                  "#00ffff", "#008080", "#0000ff", "#000080", "#ff00ff", "#800080"]
-    # Scale Radius
-    if radius == None:
-        radius = _np.abs(phasor).max()
-    # Set Tolerance
-    if tolerance == None:
-        tolerance = radius / 25
-    elif tolerance == False:
-        tolerance = -1
-    # Set Background Color
-    if bg == None:
-        bg = "#FFFFFF"
-    # Load labels if handled in other argument
-    if label != False:
-        legend = label
-    if labels != False:
-        legend = labels
-    # Check for more phasors than colors
-    numphs = len(phasor)
-    numclr = len(colors)
-    if numphs > numclr:
-        raise ValueError("ERROR: Too many phasors provided. Specify more line colors.")
-
-    if size == None:
-        # Force square figure and square axes
-        width, height = _matplotlib.rcParams['figure.figsize']
-        size = min(width, height)
-    # Make a square figure
-    fig = _plt.figure(figsize=(size, size))
-    ax = fig.add_axes([0.1, 0.1, 0.8, 0.8], polar=True, facecolor=bg)
-    _plt.grid(True)
-
-    # Plot the diagram
-    _plt.title(title + "\n")
-    handles = _np.array([])  # Empty array for plot handles
-    for i in range(numphs):
-        mag, ang_r = _c.polar(phasor[i])
-        # Plot with labels
-        if legend != False:
-            if mag > tolerance:
-                hand = _plt.arrow(0, 0, ang_r, mag, color=colors[i],
-                                  label=legend[i], linewidth=linewidth)
-            else:
-                hand = _plt.plot(0, 0, 'o', markersize=linewidth * 3,
-                                 label=legend[i], color=colors[i])
-            handles = _np.append(handles, [hand])
-        # Plot without labels
-        else:
-            _plt.arrow(0, 0, ang_r, mag, color=colors[i], linewidth=linewidth)
-    if legend != False: _plt.legend((handles), legend)
-    # Set Minimum and Maximum Radius Terms
-    ax.set_rmax(radius)
-    ax.set_rmin(0)
-    if filename != None:
-        if not any(sub in filename for sub in ['.png', '.jpg']):
-            filename += '.png'  # Add File Extension
-        _plt.savefig(filename)
-    if plot:
-        _plt.show()
-    else:
-        _plt.close()
-
-
 # Define Non-Linear Power Factor Calculator
 def nlinpf(PFtrue=False, PFdist=False, PFdisp=False):
     """
@@ -1574,7 +922,7 @@ def voltdiv(Vin, R1, R2, Rload=None):
     if (Rload == None):  # No Load Given
         Vout = Vin * R2 / (R1 + R2)
     else:  # Load was given
-        Rp = parallelz((R2, Rload))
+        Rp = R2*Rload/(R2 + Rload)
         Vout = Vin * Rp / (R1 + Rp)
     return (Vout)
 
@@ -1958,62 +1306,6 @@ def zdecompose(Zmag, XoR):
     X = R * XoR
     # Return
     return (R, X)
-
-
-# Define HP to Watts Calculation
-def hp_to_watts(hp):
-    r"""
-    Horsepower to Watts Formula.
-
-    Calculates the power (in watts) given the
-    horsepower.
-
-    .. math:: P_{\text{watts}}=P_{\text{horsepower}}\cdot745.699872
-
-    Same as `watts`.
-
-    Parameters
-    ----------
-    hp:         float
-                The horspower to compute.
-
-    Returns
-    -------
-    watts:      float
-                The power in watts.
-    """
-    return (hp * 745.699872)
-
-
-watts = hp_to_watts  # Make Duplicate Name
-
-
-# Define Watts to HP Calculation
-def watts_to_hp(watts):
-    r"""
-    Watts to Horsepower Function.
-
-    Calculates the power (in horsepower) given
-    the power in watts.
-
-    .. math:: P_{\text{horsepower}}=\frac{P_{\text{watts}}}{745.699872}
-
-    Same as `horsepower`.
-
-    Parameters
-    ----------
-    watts:      float
-                The wattage to compute.
-
-    Returns
-    -------
-    hp:         float
-                The power in horsepower.
-    """
-    return (watts / 745.699872)
-
-
-horsepower = watts_to_hp  # Make Duplicate Name
 
 
 # Define Power Reactance Calculator
@@ -2460,418 +1752,6 @@ def rtdtemp(RT, rtdtype="PT100", fahrenheit=False, Rref=None, Tref=None,
     return (temp)
 
 
-# Define Capacitor Voltage Discharge Function
-def vcapdischarge(t, Vs, R, C):
-    r"""
-    Discharging Capacitor Function.
-
-    Function to calculate the voltage of a
-    capacitor that is discharging given the time.
-
-    .. math:: V_c=V_s*e^{\frac{-t}{R*C}}
-
-    Parameters
-    ----------
-    t:          float
-                The time at which to calculate the voltage.
-    Vs:         float
-                The starting voltage for the capacitor.
-    R:          float
-                The ohmic value of the resistor being used
-                to discharge.
-    C:          float
-                Capacitive value (in Farads).
-
-    Returns
-    -------
-    Vc:         float
-                The calculated voltage of the capacitor.
-    """
-    try:
-        assert t>=0
-    except AssertionError:
-        raise ValueError("Time must be greater than or equal to zero.")
-    try:
-        assert R*C != 0
-    except AssertionError:
-        raise ValueError("Resistance and Capacitance must be non-zero.")
-    Vc = Vs * (_np.exp(-t / (R * C)))
-    return (Vc)
-
-
-# Define Capacitor Voltage Charge Function
-def vcapcharge(t, Vs, R, C):
-    r"""
-    Charging Capacitor Voltage.
-
-    Function to calculate the voltage of a
-    capacitor that is charging given the time.
-
-    .. math:: V_c=V_s*(1-e^{\frac{-t}{R*C}})
-
-    Parameters
-    ----------
-    t:          float
-                The time at which to calculate the voltage.
-    Vs:         float
-                The charging voltage for the capacitor.
-    R:          float
-                The ohmic value of the resistor being used
-                to discharge.
-    C:          float
-                Capacitive value (in Farads).
-
-    Returns
-    -------
-    Vc:         float
-                The calculated voltage of the capacitor.
-    """
-    try:
-        assert t>=0
-    except AssertionError:
-        raise ValueError("Time must be greater than or equal to zero.")
-    try:
-        assert R*C != 0
-    except AssertionError:
-        raise ValueError("Resistance and Capacitance must be non-zero.")
-    Vc = Vs * (1 - _np.exp(-t / (R * C)))
-    return (Vc)
-
-
-# Define Capacitive Energy Transfer Function
-def captransfer(t, Vs, R, Cs, Cd):
-    """
-    Capacitor Energy Transfer Function.
-
-    Calculate the voltage across a joining
-    resistor (R) that connects Cs and Cd, the
-    energy-source and -destination capacitors,
-    respectively. Calculate the final voltage
-    across both capacitors.
-
-    Parameters
-    ----------
-    t:          float
-                Time at which to calculate resistor voltage.
-    Vs:         float
-                Initial voltage across source-capacitor (Cs).
-    R:          float
-                Value of resistor that connects capacitors.
-    Cs:         float
-                Source capacitance value in Farads.
-    Cd:         float
-                Destination capacitance value in Farads.
-
-    Returns
-    -------
-    rvolt:      float
-                Voltage across the resistor at time t.
-    vfinal:     float
-                Final voltage that both capacitors settle to.
-    """
-    tau = (R * Cs * Cd) / (Cs + Cd)
-    rvolt = Vs * _np.exp(-t / tau)
-    vfinal = Vs * Cs / (Cs + Cd)
-    return (rvolt, vfinal)
-
-
-# Define Inductor Energy Formula
-def inductorenergy(L, I):
-    r"""
-    Energy Stored in Inductor Formula.
-
-    Function to calculate the energy stored in an inductor
-    given the inductance (in Henries) and the current.
-
-    .. math:: E=\frac{1}{2}*L*I^2
-
-    Parameters
-    ----------
-    L:          float
-                Inductance Value (in Henries)
-    I:          float
-                Current traveling through inductor.
-
-    Returns
-    -------
-    E:          float
-                The energy stored in the inductor (in Joules).
-    """
-    return (1 / 2 * L * I ** 2)
-
-
-# Define Inductor Charge Function
-def inductorcharge(t, Vs, R, L):
-    r"""
-    Charging Inductor Formula.
-
-    Calculates the Voltage and Current of an inductor
-    that is charging/storing energy.
-
-    .. math::
-       V_L = V_s*e^{\frac{-R*t}{L}}//
-       I_L = \frac{V_s}{R}*(1-e^{\frac{-R*t}{L}})
-
-    Parameters
-    ----------
-    t:          float
-                Time at which to calculate voltage and current.
-    Vs:         float
-                Charging voltage across inductor and resistor.
-    R:          float
-                Resistance related to inductor.
-    L:          float
-                Inductance value in Henries.
-
-    Returns
-    -------
-    Vl:         float
-                Voltage across inductor at time t.
-    Il:         float
-                Current through inductor at time t.
-    """
-    Vl = Vs * _np.exp(-R * t / L)
-    Il = Vs / R * (1 - _np.exp(-R * t / L))
-    return (Vl, Il)
-
-
-# Define Capacitive Back-to-Back Switching Formula
-def capbacktoback(C1, C2, Lm, VLN=None, VLL=None):
-    """
-    Back to Back Capacitor Transient Current Calculator.
-
-    Function to calculate the maximum current and the
-    frequency of the inrush current of two capacitors
-    connected in parallel when one (energized) capacitor
-    is switched into another (non-engergized) capacitor.
-
-    .. note:: This formula is only valid for three-phase systems.
-
-    Parameters
-    ----------
-    C1:         float
-                The capacitance of the
-    VLN:        float, exclusive
-                The line-to-neutral voltage experienced by
-                any one of the (three) capacitors in the
-                three-phase capacitor bank.
-    VLL:        float, exclusive
-                The line-to-line voltage experienced by the
-                three-phase capacitor bank.
-
-    Returns
-    -------
-    imax:       float
-                Maximum Current Magnitude during Transient
-    ifreq:      float
-                Transient current frequency
-    """
-    # Evaluate Max Current
-    imax = _np.sqrt(2 / 3) * VLL * _np.sqrt((C1 * C2) / ((C1 + C2) * Lm))
-    # Evaluate Inrush Current Frequency
-    ifreq = 1 / (2 * _np.pi * _np.sqrt(Lm * (C1 * C2) / (C1 + C2)))
-    return (imax, ifreq)
-
-
-# Define Inductor Discharge Function
-def inductordischarge(t, Io, R, L):
-    r"""
-    Discharging Inductor Formula.
-
-    Calculates the Voltage and Current of an inductor
-    that is discharging its stored energy.
-
-    .. math::
-       I_L=I_0*e^{\frac{-R*t}{L}}//
-       V_L=I_0*R*(1-e^{\frac{-R*t}{L}})
-
-    Parameters
-    ----------
-    t:          float
-                Time at which to calculate voltage and current.
-    Io:         float
-                Initial current traveling through inductor.
-    R:          float
-                Resistance being discharged to.
-    L:          float
-                Inductance value in Henries.
-
-    Returns
-    -------
-    Vl:         float
-                Voltage across inductor at time t.
-    Il:         float
-                Current through inductor at time t.
-    """
-    Il = Io * _np.exp(-R * t / L)
-    Vl = Io * R * (1 - _np.exp(-R * t / L))
-    return (Vl, Il)
-
-
-# Define Apparent Power to Farad Conversion
-def farads(VAR, V, freq=60):
-    r"""
-    Capacitance from Apparent Power Formula.
-
-    Function to calculate the required capacitance
-    in Farads to provide the desired power rating
-    (VARs).
-
-    .. math:: C = \frac{VAR}{2*\pi*freq*V^2}
-
-    Parameters
-    ----------
-    VAR:        float
-                The rated power to meet.
-    V:          float
-                The voltage across the capacitor;
-                not described as VLL or VLN, merely
-                the capacitor voltage.
-    freq:       float, optional
-                The System frequency
-
-    Returns
-    -------
-    C:          float
-                The evaluated capacitance (in Farads).
-    """
-    return (VAR / (2 * _np.pi * freq * V ** 2))
-
-
-# Define Capacitor Energy Calculation
-def capenergy(C, V):
-    r"""
-    Capacitor Energy Formula.
-
-    A simple function to calculate the stored voltage (in Joules)
-    in a capacitor with a charged voltage.
-
-    .. math:: E=\frac{1}{2}*C*V^2
-
-    Parameters
-    ----------
-    C:          float
-                Capacitance in Farads.
-    V:          float
-                Voltage across capacitor.
-
-    Returns
-    -------
-    energy:     float
-                Energy stored in capacitor (Joules).
-    """
-    energy = 1 / 2 * C * V ** 2
-    return (energy)
-
-
-# Define Capacitor Voltage Discharge Function
-def loadedvcapdischarge(t, vo, C, P):
-    # noqa: D401    "Loaded" is a valid word for this docstring
-    r"""
-    Loaded Capacitor Discharge Formula.
-
-    Returns the voltage of a discharging capacitor after time (t -
-    seconds) given initial voltage (vo - volts), capacitor size
-    (cap - Farads), and load (P - Watts).
-
-    .. math:: V_t=\sqrt{v_0^2-2*P*\frac{t}{C}}
-
-    Parameters
-    ----------
-    t:          float
-                Time at which to calculate voltage.
-    vo:         float
-                Initial capacitor voltage.
-    C:          float
-                Capacitance (in Farads)
-    P:          float
-                Load power consumption (in Watts).
-
-    Returns
-    -------
-    Vt:         float
-                Voltage of capacitor at time t.
-    """
-    Vt = _np.sqrt(vo ** 2 - 2 * P * t / C)
-    return (Vt)
-
-
-# Define Capacitor Discharge Function
-def timedischarge(Vinit, Vmin, C, P, dt=1e-3, RMS=True, Eremain=False):
-    r"""
-    Capacitor Discharge Time Formula.
-
-    Returns the time to discharge a capacitor to a specified
-    voltage given set of inputs.
-
-    Parameters
-    ----------
-    Vinit:      float
-                Initial Voltage (in volts)
-    Vmin:       float
-                Final Voltage (the minimum allowable voltage) (in volts)
-    C:          float
-                Capacitance (in Farads)
-    P:          float
-                Load Power being consumed (in Watts)
-    dt:         float, optional
-                Time step-size (in seconds) (defaults to 1e-3 | 1ms)
-    RMS:        bool, optional
-                if true converts RMS Vin to peak
-    Eremain:    bool, optional
-                if true: also returns the energy remaining in cap
-
-    Returns
-    -------
-    Returns time to discharge from Vinit to Vmin in seconds.
-    May also return remaining energy in capacitor if Eremain=True
-    """
-    t = 0  # start at time t=0
-    if RMS:
-        vo = Vinit * _np.sqrt(2)  # convert RMS to peak
-    else:
-        vo = Vinit
-    vc = loadedvcapdischarge(t, vo, C, P)  # set initial cap voltage
-    while (vc >= Vmin):
-        t = t + dt  # increment the time
-        vcp = vc  # save previous voltage
-        vc = loadedvcapdischarge(t, vo, C, P)  # calc. new voltage
-    if (Eremain):
-        E = capenergy(C, vcp)  # calc. energy
-        return (t - dt, E)
-    else:
-        return (t - dt)
-
-
-# Define Rectifier Capacitor Calculator
-def rectifiercap(Iload, fswitch, dVout):
-    r"""
-    Rectifier Capacitor Formula.
-
-    Returns the capacitance (in Farads) for a needed capacitor in
-    a rectifier configuration given the system frequency (in Hz),
-    the load (in amps) and the desired voltage ripple.
-
-    .. math:: C=\frac{I_{load}}{f_{switch}*\Delta V_{out}}
-
-    Parameters
-    ----------
-    Iload:      float
-                The load current that must be met.
-    fswitch:    float
-                The switching frequency of the system.
-    dVout:      float
-                Desired delta-V on the output.
-
-    Returns
-    -------
-    C:          float
-                Required capacitance (in Farads) to meet arguments.
-    """
-    C = Iload / (fswitch * dVout)
-    return (C)
-
-
 # Define function to find VDC setpoint
 def vscdcbus(VLL, Zs, P, Q=0, mmax=0.8, debug=False):
     """
@@ -3011,44 +1891,6 @@ def convbar(h, x, outline=True):
     _plt.show()
 
 
-# Define convolution function
-def convolve(tuple):
-    """
-    Filter Convolution Function.
-
-    Given a tuple of terms, convolves all terms in tuple to
-    return one tuple as a numpy array.
-
-    Parameters
-    ----------
-    tuple:      tuple of numpy.ndarray
-                Tuple of terms to be convolved.
-
-    Returns
-    -------
-    c:          The convolved set of the individual terms.
-                i.e. numpy.ndarray([ x1, x2, x3, ..., xn ])
-    """
-    c = sig.convolve(tuple[0], tuple[1])
-    if (len(tuple) > 2):
-        # Iterate starting with second element and continuing
-        for i in range(2, len(tuple)):
-            c = sig.convolve(c, tuple[i])
-    return (c)
-
-
-# Define Step function
-def step(t):
-    """
-    Step Function [ u(t) ].
-
-    Simple implimentation of numpy.heaviside function
-    to provide standard step-function as specified to
-    be zero at x<0, and one at x>=0.
-    """
-    return (_np.heaviside(t, 1))
-
-
 # Define Peak Calculator
 def peak(val):
     """
@@ -3073,203 +1915,6 @@ def rms(val):
     two.
     """
     return (val * _np.sqrt(0.5))
-
-
-# Arbitrary Waveform RMS Calculating Function
-def funcrms(func, T):
-    """
-    Root-Mean-Square (RMS) Evaluator for Callable Functions.
-
-    Integral-based RMS calculator, evaluates the RMS value
-    of a repetative signal (f) given the signal's specific
-    period (T)
-
-    Parameters
-    ----------
-    func:   float
-            The periodic function, a callable like f(t)
-    T:      float
-            The period of the function f, so that f(0)==f(T)
-
-    Returns
-    -------
-    RMS:    The RMS value of the function (f) over the interval ( 0, T )
-
-    """
-    fn = lambda x: func(x) ** 2
-    integral, _ = integrate(fn, 0, T)
-    return _np.sqrt(1 / T * integral)
-
-
-# Define Gaussian Function
-def gaussian(x, mu=0, sigma=1):
-    """
-    Gaussian Function.
-
-    This function is designed to generate the gaussian
-    distribution curve with configuration mu and sigma.
-
-    Parameters
-    ----------
-    x:      float
-            The input (array) x.
-    mu:     float, optional
-            Optional control argument, default=0
-    sigma:  float, optional
-            Optional control argument, default=1
-
-    Returns
-    -------
-    Computed gaussian (numpy.ndarray) of the input x
-    """
-    return (1 / (sigma * _np.sqrt(2 * _np.pi)) *
-            _np.exp(-(x - mu) ** 2 / (2 * sigma ** 2)))
-
-
-# Define Gaussian Distribution Function
-def gausdist(x, mu=0, sigma=1):
-    """
-    Gaussian Distribution Function.
-
-    This function is designed to calculate the generic
-    distribution of a gaussian function with controls
-    for mu and sigma.
-
-    Parameters
-    ----------
-    x:      numpy.ndarray
-            The input (array) x
-    mu:     float, optional
-            Optional control argument, default=0
-    sigma:  float, optional
-            Optional control argument, default=1
-
-    Returns
-    -------
-    F:      numpy.ndarray
-            Computed distribution of the gausian function at the
-            points specified by (array) x
-    """
-    F = _np.array([])
-    try:
-        lx = len(x)  # Find length of Input
-    except:
-        lx = 1  # Length 1
-        x = [x]  # Pack into list
-    for i in range(lx):
-        x_tmp = x[i]
-        # Evaluate X (altered by mu and sigma)
-        X = (x_tmp - mu) / sigma
-
-        # Define Integrand
-        def integrand(sq):
-            return (_np.exp(-sq ** 2 / 2))
-
-        integral = integrate(integrand, _np.NINF, X)  # Integrate
-        result = 1 / _np.sqrt(2 * _np.pi) * integral[0]  # Evaluate Result
-        F = _np.append(F, result)  # Append to output list
-    # Return only the 0-th value if there's only 1 value available
-    if (len(F) == 1):
-        F = F[0]
-    return (F)
-
-
-# Define Probability Density Function
-def probdensity(func, x, x0=0, scale=True):
-    """
-    Probability Density Function.
-
-    This function uses an integral to compute the probability
-    density of a given function.
-
-    Parameters
-    ----------
-    func:   function
-            The function for which to calculate the PDF
-    x:      numpy.ndarray
-            The (array of) value(s) at which to calculate
-            the PDF
-    x0:     float, optional
-            The lower-bound of the integral, starting point
-            for the PDF to be calculated over, default=0
-    scale:  bool, optional
-            The scaling to be applied to the output,
-            default=True
-
-    Returns
-    -------
-    sumx:   numpy.ndarray
-            The (array of) value(s) computed as the PDF at
-            point(s) x
-    """
-    sumx = _np.array([])
-    try:
-        lx = len(x)  # Find length of Input
-    except:
-        lx = 1  # Length 1
-        x = [x]  # Pack into list
-    # Recursively Find Probability Density
-    for i in range(lx):
-        sumx = _np.append(sumx, integrate(func, x0, x[i])[0])
-    # Return only the 0-th value if there's only 1 value available
-    if (len(sumx) == 1):
-        sumx = sumx[0]
-    else:
-        if (scale == True):
-            mx = sumx.max()
-            sumx /= mx
-        elif (scale != False):
-            sumx /= scale
-    return (sumx)
-
-
-# Define Real FFT Evaluation Function
-def rfft(arr, dt=0.01, absolute=True, resample=True):
-    """
-    RFFT Function.
-
-    This function is designed to evaluat the real FFT
-    of a input signal in the form of an array or list.
-
-    Parameters
-    ----------
-    arr:        numpy.ndarray
-                The input array representing the signal
-    dt:         float, optional
-                The time-step used for the array,
-                default=0.01
-    absolute:   bool, optional
-                Control argument to force absolute
-                values, default=True
-    resample:   bool, optional
-                Control argument specifying whether
-                the FFT output should be resampled,
-                or if it should have a specific
-                resampling rate, default=True
-
-    Returns
-    -------
-    FFT Array
-    """
-    # Calculate with Absolute Values
-    if absolute:
-        fourier = abs(_np.fft.rfft(arr))
-    else:
-        foruier = _np.fft.rfft(arr)
-    if resample == True:
-        # Evaluate the Downsampling Ratio
-        dn = int(dt * len(arr))
-        # Downsample to remove unnecessary points
-        fixedfft = filter.dnsample(fourier, dn)
-        return (fixedfft)
-    elif resample == False:
-        return (fourier)
-    else:
-        # Condition Resample Value
-        resample = int(resample)
-        # Downsample to remove unnecessary points
-        fixedfft = filter.dnsample(fourier, resample)
-        return (fixedfft)
 
 
 # Define Normalized Power Spectrum Function
@@ -3599,59 +2244,6 @@ def string_to_bits(str):
     return (data)
 
 
-# Define kWh to BTU function and vice-versa
-def kwh_to_btu(kWh):
-    r"""
-    Killo-Watt-Hours to BTU Function.
-
-    Converts kWh (killo-Watt-hours) to BTU (British Thermal Units).
-
-    .. math:: \text{BTU} = \text{kWh}\cdot3412.14
-
-    Same as `btu`.
-
-    Parameters
-    ----------
-    kWh:        float
-                The number of killo-Watt-hours
-
-    Returns
-    -------
-    BTU:        float
-                The number of British Thermal Units
-    """
-    return (kWh * 3412.14)
-
-
-btu = kwh_to_btu  # Make Duplicate Name
-
-
-def btu_to_kwh(BTU):
-    r"""
-    BTU to Killo-Watt-Hours Function.
-
-    Converts BTU (British Thermal Units) to kWh (killo-Watt-hours).
-
-    .. math:: \text{kWh} = \frac{\text{BTU}}{3412.14}
-
-    Same as `kwh`.
-
-    Parameters
-    ----------
-    BTU:        float
-                The number of British Thermal Units
-
-    Returns
-    -------
-    kWh:        float
-                The number of killo-Watt-hours
-    """
-    return (BTU / 3412.14)
-
-
-kwh = btu_to_kwh  # Make Duplicate Name
-
-
 # Define Per-Unit Impedance Formula
 def zpu(S, VLL=None, VLN=None):
     r"""
@@ -3875,176 +2467,6 @@ def geninternalv(I, Zs, Vt, Vgn=None,Zm=None, Zmp=None, Zmpp=None, Ip=None, Ipp=
     else:
         raise ValueError("Invalid Parameter Set")
     return (Ea)
-
-
-# Define Sequence Component Conversion Function
-def abc_to_seq(Mabc, reference='A'):
-    r"""
-    Phase-System to Sequence-System Conversion.
-
-    Converts phase-based values to sequence
-    components.
-
-    .. math:: M_{\text{012}}=A_{\text{012}}\cdot M_{\text{ABC}}
-
-    Same as phs_to_seq.
-
-    Parameters
-    ----------
-    Mabc:       list of complex
-                Phase-based values to be converted.
-    reference:  {'A', 'B', 'C'}
-                Single character denoting the reference,
-                default='A'
-
-    Returns
-    -------
-    M012:       numpy.ndarray
-                Sequence-based values in order of 0-1-2
-
-    See Also
-    --------
-    seq_to_abc: Sequence to Phase Conversion
-    sequencez:  Phase Impedance to Sequence Converter
-    """
-    # Condition Reference:
-    reference = reference.upper()
-    if reference == 'A':
-        M = Aabc
-    elif reference == 'B':
-        M = _np.roll(Aabc, 1, 0)
-    elif reference == 'C':
-        M = _np.roll(Aabc, 2, 0)
-    else:
-        raise ValueError("Invalid Phase Reference.")
-    return (M.dot(Mabc))
-
-
-# Define Second Name for abc_to_seq
-phs_to_seq = abc_to_seq
-
-
-# Define Phase Component Conversion Function
-def seq_to_abc(M012, reference='A'):
-    r"""
-    Sequence-System to Phase-System Conversion.
-
-    Converts sequence-based values to phase
-    components.
-
-    .. math:: M_{\text{ABC}}=A_{\text{012}}^{-1}\cdot M_{\text{012}}
-
-    Same as seq_to_phs.
-
-    Parameters
-    ----------
-    M012:       list of complex
-                Sequence-based values to convert.
-    reference:  {'A', 'B', 'C'}
-                Single character denoting the reference,
-                default='A'
-
-    Returns
-    -------
-    Mabc:       numpy.ndarray
-                Phase-based values in order of A-B-C
-
-    See Also
-    --------
-    abc_to_seq: Phase to Sequence Conversion
-    sequencez:  Phase Impedance to Sequence Converter
-    """
-    # Compute Dot Product
-    M = A012.dot(M012)
-    # Condition Reference:
-    reference = reference.upper()
-    if reference == 'A':
-        pass
-    elif reference == 'B':
-        M = _np.roll(M, 1, 0)
-    elif reference == 'C':
-        M = _np.roll(M, 2, 0)
-    else:
-        raise ValueError("Invalid Phase Reference.")
-    return (M)
-
-
-# Define Second Name for seq_to_abc
-seq_to_phs = seq_to_abc
-
-
-# Define Sequence Impedance Calculator
-def sequencez(Zabc, reference='A', resolve=False, diag=False, round=3):
-    r"""
-    Sequence Impedance Calculator.
-
-    Accepts the phase (ABC-domain) impedances for a
-    system and calculates the sequence (012-domain)
-    impedances for the same system. If the argument
-    `resolve` is set to true, the function will
-    combine terms into the set of [Z0, Z1, Z2].
-
-    When resolve is False:
-
-    .. math:: Z_{\text{012-M}}=A_{\text{012}}^{-1}Z_{\text{ABC}}A_{\text{012}}
-
-    When resolve is True:
-
-    .. math:: Z_{\text{012}}=A_{\text{012}}Z_{\text{ABC}}A_{\text{012}}^{-1}
-
-    Parameters
-    ----------
-    Zabc:       numpy.ndarray of complex
-                2-D (3x3) matrix of complex values
-                representing the phasor impedances
-                in the ABC-domain.
-    reference:  {'A', 'B', 'C'}
-                Single character denoting the reference,
-                default='A'
-    resolve:    bool, optional
-                Control argument to force the function to
-                evaluate the individual sequence impedances
-                [Z0, Z1, Z2], default=False
-    diag:       bool, optional
-                Control argument to force the function to
-                reduce the matrix to its diagonal terms.
-    round:      int, optional
-                Integer denoting number of decimal places
-                resulting matrix should be rounded to.
-                default=3
-
-    Returns
-    -------
-    Z012:       numpy.ndarray of complex
-                2-D (3x3) matrix of complex values
-                representing the sequence impedances
-                in the 012-domain
-
-    See Also
-    --------
-    seq_to_abc: Sequence to Phase Conversion
-    abc_to_seq: Phase to Sequence Conversion
-    """
-    # Condition Reference
-    reference = reference.upper()
-    rollrate = {'A': 0, 'B': 1, 'C': 2}
-    # Test Validity
-    if reference not in rollrate:
-        raise ValueError("Invalad Phase Reference")
-    # Determine Roll Factor
-    roll = rollrate[reference]
-    # Evaluate Matricies
-    M012 = _np.roll(A012, roll, 0)
-    Minv = _np.linalg.inv(M012)
-    # Compute Sequence Impedances
-    if resolve:
-        Z012 = M012.dot(Zabc.dot(Minv))
-    else:
-        Z012 = Minv.dot(Zabc.dot(M012))
-    # Reduce to Diagonal Terms if Needed
-    if diag:
-        Z012 = [Z012[0][0], Z012[1][1], Z012[2][2]]
-    return (_np.around(Z012, round))
 
 
 # FFT Coefficient Calculator Function
@@ -4968,62 +3390,6 @@ def xfmphs(style="DY", shift=30):
     phase = _np.exp(1j * _np.radians(v * abs(shift)))
     # Return
     return (phase)
-
-
-# Define Simple Radians to Hertz Converter
-def rad_to_hz(radians):
-    r"""
-    Radians to Hertz Converter.
-
-    Accepts a frequency in radians/sec and calculates
-    the hertzian frequency (in Hz).
-
-    .. math:: f_{\text{Hz}} = \frac{f_{\text{rad/sec}}}{2\cdot\pi}
-
-    Same as `hertz`.
-
-    Parameters
-    ----------
-    radians:    float
-                The frequency (represented in radians/sec)
-
-    Returns
-    -------
-    hertz:      float
-                The frequency (represented in Hertz)
-    """
-    return (radians / (2 * _np.pi))  # Evaluate and Return
-
-
-hertz = rad_to_hz  # Make Duplicate Name
-
-
-# Define Simple Hertz to Radians Converter
-def hz_to_rad(hertz):
-    r"""
-    Hertz to Radians Converter.
-
-    Accepts a frequency in Hertz and calculates
-    the frequency in radians/sec.
-
-    .. math:: f_{\text{rad/sec}} = f_{\text{Hz}}\cdot2\cdot\pi
-
-    Same as `radsec`.
-
-    Parameters
-    ----------
-    hertz:      float
-                The frequency (represented in Hertz)
-
-    Returns
-    -------
-    radians:    float
-                The frequency (represented in radians/sec)
-    """
-    return (hertz * (2 * _np.pi))  # Evaluate and Return
-
-
-radsec = hz_to_rad  # Make Duplicate Name
 
 
 # Define Induction Machine Thevenin Voltage Calculator
@@ -6404,94 +4770,6 @@ def vipf(V=None, I=None, PF=1, find=''):
         return (PF)
     else:
         return (V, I, PF)
-
-
-# Define Angular Velocity Conversion Functions
-def rad_to_rpm(rad):
-    """
-    Radians-per-Second to RPM Converter.
-
-    Given the angular velocity in rad/sec, this function will evaluate the
-    velocity in RPM (Revolutions-Per-Minute).
-
-    Parameters
-    ----------
-    rad:        float
-                The angular velocity in radians-per-second
-
-    Returns
-    -------
-    rpm:        float
-                The angular velocity in revolutions-per-minute (RPM)
-    """
-    rpm = 60 / (2 * _np.pi) * rad
-    return (rpm)
-
-
-# Define Angular Velocity Conversion Functions
-def rpm_to_rad(rpm):
-    """
-    RPM to Radians-per-Second Converter.
-
-    Given the angular velocity in RPM (Revolutions-Per-Minute), this function
-    will evaluate the velocity in rad/sec.
-
-    Parameters
-    ----------
-    rpm:        float
-                The angular velocity in revolutions-per-minute (RPM)
-
-    Returns
-    -------
-    rad:        float
-                The angular velocity in radians-per-second
-    """
-    rad = 2 * _np.pi / 60 * rpm
-    return (rad)
-
-
-# Define Angular Velocity Conversion Functions
-def hz_to_rpm(hz):
-    """
-    Hertz to RPM Converter.
-
-    Given the angular velocity in Hertz, this function will evaluate the
-    velocity in RPM (Revolutions-Per-Minute).
-
-    Parameters
-    ----------
-    hz:         float
-                The angular velocity in Hertz
-
-    Returns
-    -------
-    rpm:        float
-                The angular velocity in revolutions-per-minute (RPM)
-    """
-    rpm = hz * 60
-    return (rpm)
-
-
-# Define Angular Velocity Conversion Functions
-def rpm_to_hz(rpm):
-    """
-    RPM to Hertz Converter.
-
-    Given the angular velocity in RPM (Revolutions-Per-Minute), this function
-    will evaluate the velocity in Hertz.
-
-    Parameters
-    ----------
-    rpm:        float
-                The angular velocity in revolutions-per-minute (RPM)
-
-    Returns
-    -------
-    hz:         float
-                The angular velocity in Hertz
-    """
-    hz = rpm / 60
-    return (hz)
 
 
 # Define Synchronous Speed Calculator
