+++ electricpy/phasor.py
@@ -16,6 +16,8 @@ to aid in scientific calculations.
 # Import Required Packages
 import numpy as _np
 import cmath as _c
+import matplotlib as _matplotlib
+import matplotlib.pyplot as _plt
 
 
 # Define Phase Angle Generator
@@ -352,4 +354,116 @@ def parallelz(*args):
             for i in range(2, L):
                 Zp = (1 / Zp + 1 / Z[i]) ** (-1)
     return (Zp)
+
+# Define Phasor Plot Generator
+def phasorplot(phasor, title="Phasor Diagram", legend=False, bg=None,
+            colors=None, radius=None, linewidth=None, size=None,
+            filename=None, plot=True, label=False, labels=False,
+            tolerance=None):
+    """
+    Phasor Plotting Function.
+
+    This function is designed to plot a phasor-diagram with angles in degrees
+    for up to 12 phasor sets. Phasors must be passed as a complex number set,
+    (e.g. [ m+ja, m+ja, m+ja, ... , m+ja ] ).
+
+    Parameters
+    ----------
+    phasor:     list of complex
+                The set of phasors to be plotted.
+    title:      string, optional
+                The Plot Title, default="Phasor Diagram"
+    legend:     bool, optional
+                Control argument to enable displaying the legend, must be passed
+                as an array or list of strings. `label` and `labels` are mimic-
+                arguments and will perform similar operation, default=False
+    bg:         string, optional
+                Background-Color control, default="#d5de9c"
+    radius:     float, optional
+                The diagram radius, unless specified, automatically scales
+    colors:     list of str, optional
+                List of hexidecimal color strings denoting the line colors to use.
+    filename:   string, optional
+                String of filename, if set, will force function to save image.
+    plot:       bool, optional
+                Control argument to disable plotting. default=True
+    size:       float, optional
+                Control argument for figure size. default=None
+    linewidth:  float, optional
+                Control argument to declare the line thickness. default=None
+    tolerance:  float, optional
+                Minimum magnitude to plot, anything less than tolerance will be
+                plotted as a single point at the origin, by default, the tolerance
+                is scaled to be 1/25-th the maximum radius. To disable the tolerance,
+                simply provide either False or -1.
+    """
+    # Load Complex Values if Necessary
+    try:
+        len(phasor)
+    except TypeError:
+        phasor = [phasor]
+    # Manage Colors
+    if colors == None:
+        colors = ["#FF0000", "#800000", "#FFFF00", "#808000", "#00ff00", "#008000",
+                "#00ffff", "#008080", "#0000ff", "#000080", "#ff00ff", "#800080"]
+    # Scale Radius
+    if radius == None:
+        radius = _np.abs(phasor).max()
+    # Set Tolerance
+    if tolerance == None:
+        tolerance = radius / 25
+    elif tolerance == False:
+        tolerance = -1
+    # Set Background Color
+    if bg == None:
+        bg = "#FFFFFF"
+    # Load labels if handled in other argument
+    if label != False:
+        legend = label
+    if labels != False:
+        legend = labels
+    # Check for more phasors than colors
+    numphs = len(phasor)
+    numclr = len(colors)
+    if numphs > numclr:
+        raise ValueError("ERROR: Too many phasors provided. Specify more line colors.")
+
+    if size == None:
+        # Force square figure and square axes
+        width, height = _matplotlib.rcParams['figure.figsize']
+        size = min(width, height)
+    # Make a square figure
+    fig = _plt.figure(figsize=(size, size))
+    ax = fig.add_axes([0.1, 0.1, 0.8, 0.8], polar=True, facecolor=bg)
+    _plt.grid(True)
+
+    # Plot the diagram
+    _plt.title(title + "\n")
+    handles = _np.array([])  # Empty array for plot handles
+    for i in range(numphs):
+        mag, ang_r = _c.polar(phasor[i])
+        # Plot with labels
+        if legend != False:
+            if mag > tolerance:
+                hand = _plt.arrow(0, 0, ang_r, mag, color=colors[i],
+                                label=legend[i], linewidth=linewidth)
+            else:
+                hand = _plt.plot(0, 0, 'o', markersize=linewidth * 3,
+                                label=legend[i], color=colors[i])
+            handles = _np.append(handles, [hand])
+        # Plot without labels
+        else:
+            _plt.arrow(0, 0, ang_r, mag, color=colors[i], linewidth=linewidth)
+    if legend != False: _plt.legend((handles), legend)
+    # Set Minimum and Maximum Radius Terms
+    ax.set_rmax(radius)
+    ax.set_rmin(0)
+    if filename != None:
+        if not any(sub in filename for sub in ['.png', '.jpg']):
+            filename += '.png'  # Add File Extension
+        _plt.savefig(filename)
+    if plot:
+        _plt.show()
+    else:
+        _plt.close()
 # END
